+++ src/binding/bpm.rs
@@ -1,15 +1,16 @@
 extern crate alloc;
 
 use super::*;
+use crate::Float;
 use core::ops::Deref;
 use spin::Mutex;
 
 pub trait Clock: Send {
-    fn bpm(&self) -> f32;
-    fn set_bpm(&mut self, bpm: f32);
+    fn bpm(&self) -> Float;
+    fn set_bpm(&mut self, bpm: Float);
 
-    fn period_micros(&self) -> f32;
-    fn set_period_micros(&mut self, period_micros: f32);
+    fn period_micros(&self) -> Float;
+    fn set_period_micros(&mut self, period_micros: Float);
 
     fn ppq(&self) -> usize;
     fn set_ppq(&mut self, ppq: usize);
@@ -17,14 +18,14 @@ pub trait Clock: Send {
 
 #[derive(Debug, Copy, Clone)]
 pub struct ClockData {
-    pub bpm: f32,
-    pub period_micros: f32,
+    pub bpm: Float,
+    pub period_micros: Float,
     pub ppq: usize,
 }
 
 macro_rules! period_micro {
     ($bpm:expr, $ppq:expr) => {
-        60e6f32 / ($bpm * $ppq as f32)
+        60.0e6 / ($bpm * $ppq as Float)
     };
 }
 
@@ -47,11 +48,11 @@ pub struct ClockBPMBinding<T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + C
 pub struct ClockPPQBinding<T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + Clone>(pub T);
 
 impl ClockData {
-    pub fn period_micro(bpm: f32, ppq: usize) -> f32 {
+    pub fn period_micro(bpm: Float, ppq: usize) -> Float {
         period_micro!(bpm, ppq)
     }
 
-    pub fn new(bpm: f32, ppq: usize) -> Self {
+    pub fn new(bpm: Float, ppq: usize) -> Self {
         Self {
             bpm,
             period_micros: Self::period_micro(bpm, ppq),
@@ -73,26 +74,26 @@ impl Default for ClockData {
 }
 
 impl Clock for ClockData {
-    fn bpm(&self) -> f32 {
+    fn bpm(&self) -> Float {
         self.bpm
     }
 
-    fn set_bpm(&mut self, bpm: f32) {
-        self.bpm = if bpm < 0f32 { 0.001f32 } else { bpm };
+    fn set_bpm(&mut self, bpm: Float) {
+        self.bpm = if bpm < 0.0 { 0.001 } else { bpm };
         self.period_micros = Self::period_micro(self.bpm, self.ppq);
     }
 
-    fn period_micros(&self) -> f32 {
+    fn period_micros(&self) -> Float {
         self.period_micros
     }
 
-    fn set_period_micros(&mut self, period_micros: f32) {
-        self.period_micros = if period_micros < 0.001f32 {
-            0.001f32
+    fn set_period_micros(&mut self, period_micros: Float) {
+        self.period_micros = if period_micros < 0.001 {
+            0.001
         } else {
             period_micros
         };
-        self.bpm = 60e6f32 / (self.period_micros * self.ppq as f32);
+        self.bpm = 60.0e6 / (self.period_micros * self.ppq as Float);
     }
 
     fn ppq(&self) -> usize {
@@ -132,38 +133,38 @@ where
     }
 }
 
-impl<T> ParamBindingSet<f32> for ClockPeriodMicroBinding<T>
+impl<T> ParamBindingSet<Float> for ClockPeriodMicroBinding<T>
 where
     T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + Clone,
 {
-    fn set(&self, value: f32) {
+    fn set(&self, value: Float) {
         self.0.lock().set_period_micros(value);
     }
 }
 
-impl<T> ParamBindingGet<f32> for ClockPeriodMicroBinding<T>
+impl<T> ParamBindingGet<Float> for ClockPeriodMicroBinding<T>
 where
     T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + Clone,
 {
-    fn get(&self) -> f32 {
+    fn get(&self) -> Float {
         self.0.lock().period_micros()
     }
 }
 
-impl<T> ParamBindingSet<f32> for ClockBPMBinding<T>
+impl<T> ParamBindingSet<Float> for ClockBPMBinding<T>
 where
     T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + Clone,
 {
-    fn set(&self, value: f32) {
+    fn set(&self, value: Float) {
         self.0.lock().set_bpm(value);
     }
 }
 
-impl<T> ParamBindingGet<f32> for ClockBPMBinding<T>
+impl<T> ParamBindingGet<Float> for ClockBPMBinding<T>
 where
     T: Deref<Target = Mutex<dyn Clock>> + Sync + Send + Clone,
 {
-    fn get(&self) -> f32 {
+    fn get(&self) -> Float {
         self.0.lock().bpm()
     }
 }
@@ -196,29 +197,29 @@ mod tests {
 
     #[test]
     fn bpm_value_test() {
-        assert_eq!(5208f32, bpm::ClockData::period_micro(120.0, 96).floor());
-        assert_eq!(20833f32, bpm::ClockData::period_micro(120.0, 24).floor());
+        assert_eq!(5208f64, bpm::ClockData::period_micro(120.0, 96).floor());
+        assert_eq!(20833f64, bpm::ClockData::period_micro(120.0, 24).floor());
 
         let mut c = bpm::ClockData::new(120.0, 96);
-        assert_eq!(5208f32, c.period_micros().floor());
-        assert_eq!(120f32, c.bpm());
+        assert_eq!(5208f64, c.period_micros().floor());
+        assert_eq!(120f64, c.bpm());
         assert_eq!(96, c.ppq());
 
         c.set_ppq(24);
-        assert_eq!(20833f32, c.period_micros().floor());
-        assert_eq!(120f32, c.bpm());
+        assert_eq!(20833f64, c.period_micros().floor());
+        assert_eq!(120f64, c.bpm());
         assert_eq!(24, c.ppq());
 
         c.set_bpm(2.0);
         c.set_ppq(96);
-        assert_eq!(2f32, c.bpm());
+        assert_eq!(2f64, c.bpm());
         assert_eq!(96, c.ppq());
-        assert_ne!(5208f32, c.period_micros().floor());
+        assert_ne!(5208f64, c.period_micros().floor());
 
-        c.set_period_micros(5_208.333333f32);
-        assert_eq!(120f32, c.bpm().floor());
+        c.set_period_micros(5_208.333333f64);
+        assert_eq!(120f64, c.bpm().floor());
         assert_eq!(96, c.ppq());
-        assert_eq!(5208f32, c.period_micros().floor());
+        assert_eq!(5208f64, c.period_micros().floor());
     }
 
     #[test]
@@ -231,40 +232,40 @@ mod tests {
         let micros2 = micros.clone();
 
         let c = b.clone();
-        assert_eq!(5208f32, c.lock().period_micros().floor());
-        assert_eq!(5208f32, micros.get().floor());
-        assert_eq!(120f32, c.lock().bpm());
-        assert_eq!(120f32, bpm.get());
+        assert_eq!(5208f64, c.lock().period_micros().floor());
+        assert_eq!(5208f64, micros.get().floor());
+        assert_eq!(120f64, c.lock().bpm());
+        assert_eq!(120f64, bpm.get());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
 
         ppq.set(24);
-        assert_eq!(20833f32, c.lock().period_micros().floor());
-        assert_eq!(20833f32, micros.get().floor());
-        assert_eq!(120f32, c.lock().bpm());
-        assert_eq!(120f32, bpm.get());
+        assert_eq!(20833f64, c.lock().period_micros().floor());
+        assert_eq!(20833f64, micros.get().floor());
+        assert_eq!(120f64, c.lock().bpm());
+        assert_eq!(120f64, bpm.get());
         assert_eq!(24, c.lock().ppq());
         assert_eq!(24, ppq.get());
 
         bpm.set(2.0);
         ppq.set(96);
-        assert_eq!(2f32, c.lock().bpm());
-        assert_eq!(2f32, bpm.get());
+        assert_eq!(2f64, c.lock().bpm());
+        assert_eq!(2f64, bpm.get());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
-        assert_ne!(5208f32, c.lock().period_micros().floor());
-        assert_ne!(5208f32, micros.get().floor());
+        assert_ne!(5208f64, c.lock().period_micros().floor());
+        assert_ne!(5208f64, micros.get().floor());
 
-        micros2.set(5_208.333333f32);
-        assert_eq!(120f32, c.lock().bpm().floor());
-        assert_eq!(120f32, bpm.get().floor());
+        micros2.set(5_208.333333f64);
+        assert_eq!(120f64, c.lock().bpm().floor());
+        assert_eq!(120f64, bpm.get().floor());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
-        assert_eq!(5208f32, c.lock().period_micros().floor());
-        assert_eq!(5208f32, micros.get().floor());
+        assert_eq!(5208f64, c.lock().period_micros().floor());
+        assert_eq!(5208f64, micros.get().floor());
     }
 
-    static CLOCK: Mutex<bpm::ClockData> = Mutex::new(make_clock!(120f32, 96));
+    static CLOCK: Mutex<bpm::ClockData> = Mutex::new(make_clock!(120f64, 96));
     #[test]
     fn bpm_binding_static_test() {
         let b = &CLOCK as &'static Mutex<dyn Clock>;
@@ -274,36 +275,36 @@ mod tests {
         let micros2 = micros.clone();
 
         let c = b.clone();
-        assert_eq!(5208f32, c.lock().period_micros().floor());
-        assert_eq!(5208f32, micros.get().floor());
-        assert_eq!(120f32, c.lock().bpm());
-        assert_eq!(120f32, bpm.get());
+        assert_eq!(5208f64, c.lock().period_micros().floor());
+        assert_eq!(5208f64, micros.get().floor());
+        assert_eq!(120f64, c.lock().bpm());
+        assert_eq!(120f64, bpm.get());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
 
         ppq.set(24);
-        assert_eq!(20833f32, c.lock().period_micros().floor());
-        assert_eq!(20833f32, micros.get().floor());
-        assert_eq!(120f32, c.lock().bpm());
-        assert_eq!(120f32, bpm.get());
+        assert_eq!(20833f64, c.lock().period_micros().floor());
+        assert_eq!(20833f64, micros.get().floor());
+        assert_eq!(120f64, c.lock().bpm());
+        assert_eq!(120f64, bpm.get());
         assert_eq!(24, c.lock().ppq());
         assert_eq!(24, ppq.get());
 
         bpm.set(2.0);
         ppq.set(96);
-        assert_eq!(2f32, c.lock().bpm());
-        assert_eq!(2f32, bpm.get());
+        assert_eq!(2f64, c.lock().bpm());
+        assert_eq!(2f64, bpm.get());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
-        assert_ne!(5208f32, c.lock().period_micros().floor());
-        assert_ne!(5208f32, micros.get().floor());
+        assert_ne!(5208f64, c.lock().period_micros().floor());
+        assert_ne!(5208f64, micros.get().floor());
 
-        micros2.set(5_208.333333f32);
-        assert_eq!(120f32, c.lock().bpm().floor());
-        assert_eq!(120f32, bpm.get().floor());
+        micros2.set(5_208.333333f64);
+        assert_eq!(120f64, c.lock().bpm().floor());
+        assert_eq!(120f64, bpm.get().floor());
         assert_eq!(96, c.lock().ppq());
         assert_eq!(96, ppq.get());
-        assert_eq!(5208f32, c.lock().period_micros().floor());
-        assert_eq!(5208f32, micros.get().floor());
+        assert_eq!(5208f64, c.lock().period_micros().floor());
+        assert_eq!(5208f64, micros.get().floor());
     }
 }
